+++ reader_service/internal/repository/product/mongo_repository_impl.go
@@ -6,6 +6,7 @@ import (
 	"github.com/minhwalker/cqrs-microservices/pkg/logger"
 	"github.com/minhwalker/cqrs-microservices/pkg/utils"
 	"github.com/minhwalker/cqrs-microservices/reader_service/config"
+	"github.com/minhwalker/cqrs-microservices/repository"
 	"github.com/opentracing/opentracing-go"
 	"github.com/pkg/errors"
 	uuid "github.com/satori/go.uuid"
@@ -21,7 +22,7 @@ type mongoRepository struct {
 	db  *mongo.Client
 }
 
-func NewMongoRepository(log logger.Logger, cfg *config.Config, db *mongo.Client) Repository {
+func NewMongoRepository(log logger.Logger, cfg *config.Config, db *mongo.Client) repository.Repository {
 	return &mongoRepository{log: log, cfg: cfg, db: db}
 }
