+++ reader_service/internal/repository/product/pg_repository_impl.go
@@ -2,13 +2,17 @@ package product
 
 import (
 	"context"
+	"fmt"
 	"github.com/jackc/pgx/v4/pgxpool"
 	"github.com/minhwalker/cqrs-microservices/models"
 	"github.com/minhwalker/cqrs-microservices/pkg/logger"
-	"github.com/minhwalker/cqrs-microservices/writer_service/config"
+	"github.com/minhwalker/cqrs-microservices/pkg/utils"
+	"github.com/minhwalker/cqrs-microservices/reader_service/config"
+	"github.com/minhwalker/cqrs-microservices/repository"
 	"github.com/opentracing/opentracing-go"
 	"github.com/pkg/errors"
 	uuid "github.com/satori/go.uuid"
+	"strings"
 )
 
 type productRepository struct {
@@ -17,7 +21,7 @@ type productRepository struct {
 	db  *pgxpool.Pool
 }
 
-func NewProductRepository(log logger.Logger, cfg *config.Config, db *pgxpool.Pool) *productRepository {
+func NewProductRepository(log logger.Logger, cfg *config.Config, db *pgxpool.Pool) repository.Repository {
 	return &productRepository{log: log, cfg: cfg, db: db}
 }
 
@@ -78,7 +82,7 @@ func (p *productRepository) GetProductById(ctx context.Context, uuid uuid.UUID) 
 	return &product, nil
 }
 
-func (p *productRepository) DeleteProductByID(ctx context.Context, uuid uuid.UUID) error {
+func (p *productRepository) DeleteProduct(ctx context.Context, uuid uuid.UUID) error {
 	span, ctx := opentracing.StartSpanFromContext(ctx, "productRepository.DeleteProductByID")
 	defer span.Finish()
 
@@ -88,4 +92,63 @@ func (p *productRepository) DeleteProductByID(ctx context.Context, uuid uuid.UUI
 	}
 
 	return nil
+}
+
+func (p *productRepository) Search(ctx context.Context, search string, pagination *utils.Pagination) (*models.ProductsList, error) {
+	span, ctx := opentracing.StartSpanFromContext(ctx, "postgresRepository.Search")
+	defer span.Finish()
+
+	query := `
+		SELECT product_id, name, description, price 
+		FROM products
+		WHERE name ILIKE $1 OR description ILIKE $1
+		LIMIT $2 OFFSET $3
+	`
+
+	// Construct the search query to perform case-insensitive search using ILIKE
+	searchTerm := fmt.Sprintf("%%%s%%", strings.ToLower(search))
+
+	rows, err := p.db.Query(ctx, query, searchTerm, pagination.GetLimit(), pagination.GetOffset())
+	if err != nil {
+		p.traceErr(span, err)
+		return nil, errors.Wrap(err, "Query")
+	}
+	defer rows.Close()
+
+	products := make([]*models.Product, 0)
+
+	for rows.Next() {
+		var prod models.Product
+		if err := rows.Scan(&prod.ProductID, &prod.Name, &prod.Description, &prod.Price); err != nil {
+			p.traceErr(span, err)
+			return nil, errors.Wrap(err, "Scan")
+		}
+		products = append(products, &prod)
+	}
+
+	if err := rows.Err(); err != nil {
+		span.SetTag("error", true)
+		span.LogKV("error_code", err.Error())
+		return nil, errors.Wrap(err, "rows.Err")
+	}
+
+	countQuery := `
+		SELECT COUNT(*) 
+		FROM products
+		WHERE name ILIKE $1 OR description ILIKE $1
+	`
+
+	var count int64
+	err = p.db.QueryRow(ctx, countQuery, searchTerm).Scan(&count)
+	if err != nil {
+		p.traceErr(span, err)
+		return nil, errors.Wrap(err, "QueryRow")
+	}
+
+	return models.NewProductListWithPagination(products, count, pagination), nil
+}
+
+func (p *productRepository) traceErr(span opentracing.Span, err error) {
+	span.SetTag("error", true)
+	span.LogKV("error_code", err.Error())
 }
