--- writer_service/internal/repository/product/repository.go
@@ -1,15 +1,0 @@
-package product
-
-import (
-	"context"
-	"github.com/minhwalker/cqrs-microservices/models"
-	uuid "github.com/satori/go.uuid"
-)
-
-type Repository interface {
-	CreateProduct(ctx context.Context, product *models.Product) (*models.Product, error)
-	UpdateProduct(ctx context.Context, product *models.Product) (*models.Product, error)
-	DeleteProductByID(ctx context.Context, uuid uuid.UUID) error
-
-	GetProductById(ctx context.Context, uuid uuid.UUID) (*models.Product, error)
-}
