+++ writer_service/internal/services/product/service.go
@@ -3,8 +3,8 @@ package product
 import (
 	kafkaClient "github.com/minhwalker/cqrs-microservices/pkg/kafka"
 	"github.com/minhwalker/cqrs-microservices/pkg/logger"
+	"github.com/minhwalker/cqrs-microservices/repository"
 	"github.com/minhwalker/cqrs-microservices/writer_service/config"
-	"github.com/minhwalker/cqrs-microservices/writer_service/internal/repository/product"
 	commands2 "github.com/minhwalker/cqrs-microservices/writer_service/internal/services/product/commands"
 	queries2 "github.com/minhwalker/cqrs-microservices/writer_service/internal/services/product/queries"
 )
@@ -14,7 +14,7 @@ type ProductService struct {
 	Queries  *queries2.ProductQueries
 }
 
-func NewProductService(log logger.Logger, cfg *config.Config, pgRepo product.Repository, kafkaProducer kafkaClient.Producer) *ProductService {
+func NewProductService(log logger.Logger, cfg *config.Config, pgRepo repository.RepositoryWriter, kafkaProducer kafkaClient.Producer) *ProductService {
 
 	updateProductHandler := commands2.NewUpdateProductHandler(log, cfg, pgRepo, kafkaProducer)
 	createProductHandler := commands2.NewCreateProductHandler(log, cfg, pgRepo, kafkaProducer)
